% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\docType{data}
\name{cats}
\alias{cats}
\title{Data on cat visits to Austin vets}
\format{A data frame with 29421 rows and 37 variables:
\describe{
  \item{age_upon_outcome}{age at outcome}
  \item{animal_id}{Animal ID}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
  \item{colname}{brief description}
}}
\usage{
cats
}
\description{
Write some words that cover 2-4 lines worth of
information that say something interesting about
the data that are collected.
}
\keyword{datasets}
