% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percent.R
\name{get_percent}
\alias{get_percent}
\title{Get a percentage}
\usage{
get_percent(x)
}
\arguments{
\item{a}{A percentage, rounded to one space}
}
\value{
The percent of \code{a given number, case, etc}
}
\description{
Get a percentage
}
\examples{
get_percent(a given number from data set)
}
