% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_cat.R
\name{graph_cat}
\alias{graph_cat}
\title{Create bar graph from results of name_cat function}
\usage{
graph_cat(df)
}
\arguments{
\item{df}{a dataframe from the name_cat function}
}
\value{
a bar graph showing frequency of names
}
\description{
Create bar graph from results of name_cat function
}
\examples{
graph_cat(name_cat("nut"))
}
