% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cat.R
\name{name_cat}
\alias{name_cat}
\title{Return number of cats with a given name or partial name}
\usage{
name_cat(x)
}
\arguments{
\item{x}{the name or partial name of a cat}
}
\value{
a dataframe with the matching names and number of cats
}
\description{
Return number of cats with a given name or partial name
}
\examples{
name_cat("nut") #returns cat with "nut" in their names
}
