% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_cat.R
\name{top_cat}
\alias{top_cat}
\title{Return top N cat names}
\usage{
top_cat(N)
}
\arguments{
\item{N}{the number of names to return}
}
\value{
a dataframe with the top N names and number of cats
}
\description{
Return top N cat names
}
\examples{
top_cat(20) #returns top 20 cat names
}
